Free Shareable reference text-book
For a deeper dive into the chapters, please refer to the text-book. The text book made available under a Creative  Commons Attribution 4.0 License. If you make a change, feel free to  add your name and organization to the list of contributors on this  page as you republish the materials by contacting the Author of the book